Wordlists, InteractivePanel, and RevealByColor
I've added two new wordlists containing many more words than the one we were using. The issue is that the paring function needs to be optimized and there is an odd visual bug with the larger word lists.

InteractivePanel is a superclass of Guess and Reveal panels which contains similar elements from both.

RevealByColor is only in its infancy. It currently only has extra buttons with no functionality.