More Difficulty Data and Small Tweaks
Had someone play the game to gather more difficulty data.

Changed how modular difficulty works to differentiate between winning on the final guess and losing on the final guess. That is, I use guessesMade-1 for the calculation. This means that the data I gathered before was not using this metric and is therefore techincally skewed but it is a small change that wouldn't affect the difficulty of the image greatly,